Find and fix several bugs, one critical
+ Fix the Block:max_wavespeed function: it was sending self._block
  rather than b._block to the C function

+ Patch potential memory errors due to filling into an unallocated
  parent

+ Change the data type of getparami from long to int: Lua code was
  incorrectly using long

+++ modules/FishSim.lua
@@ -63,7 +63,7 @@ end
 
 function FishSimulation:report_configuration()
    local scheme = self.scheme
-   local enum = array.vector(1, 'long')
+   local enum = array.vector(1, 'int')
    local cfg = { }
    for _,k in pairs{'RIEMANN_SOLVER',
 		    'RECONSTRUCTION',
